requirements.txt
Flask==3.0.3
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
